Split v24_2 champion_challenger into a multi-module package
The v24_2 `champion_challenger.py` (~770 lines) was the largest
remaining single file in the versioned API client after #732 split
`prediction.py` and `prediction_studio.py`. Apply the same
mixin/sync/async layout used there:

- `__init__.py`  — re-exports the public surface
- `_mixin.py`    — `_ChampionChallengerV24_2Mixin`
- `_sync.py`     — `ChampionChallenger`
- `_async.py`    — `AsyncChampionChallenger`

Pure mechanical split: no public API, signature, or behaviour changes.
External imports (`from ...v24_2.champion_challenger import
ChampionChallenger`, etc.) continue to work unchanged.
