+++ python/pdstools/infinity/resources/prediction_studio/v24_2/champion_challenger/__init__.py
@@ -1,0 +1,9 @@
+from ._async import AsyncChampionChallenger
+from ._mixin import _ChampionChallengerV24_2Mixin
+from ._sync import ChampionChallenger
+
+__all__ = [
+    "AsyncChampionChallenger",
+    "ChampionChallenger",
+    "_ChampionChallengerV24_2Mixin",
+]
